re-organize product filters tools
