+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/status-filter/inspector.tsx
@@ -21,7 +21,7 @@ import {
  * Internal dependencies
  */
 import { BlockAttributes, EditProps } from './types';
-import { getInnerBlockByName } from '../../utils';
+import { getInnerBlockByName } from '../../utils/block';
 
 let displayStyleOptions: Block[] = [];
