+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/utils/block.ts
